support core v4.2

+++ plugin_info/install.php
@@ -19,51 +19,49 @@
 require_once dirname(__FILE__) . '/../../../core/php/core.inc.php';
 
 function qivivo_install() {
-    config::save('functionality::cron5::enable', 0, 'qivivo');
-    config::save('functionality::cron15::enable', 1, 'qivivo');
-    config::save('pluginversion', 2.1, 'qivivo');
+	config::save('functionality::cron5::enable', 0, 'qivivo');
+	config::save('functionality::cron15::enable', 1, 'qivivo');
+	config::save('pluginversion', 2.1, 'qivivo');
 }
 
 function qivivo_update() {
-    //New v2 version:
-    $pluginVersion = config::byKey('pluginversion', 'qivivo');
-    if ($pluginVersion == '')
-    {
-        $pluginVersion = 1.9;
-    }
+	//New v2 version:
+	$pluginVersion = config::byKey('pluginversion', 'qivivo');
+	if ($pluginVersion == '') {
+		$pluginVersion = 1.9;
+	}
 
-    if ($pluginVersion < 2.0)
-    {
-        //new custom API for new Comap interface:
-        $folderPath = dirname(__FILE__) . '/../../qivivo/exportedPrograms/';
-        if (is_dir($folderPath))
-        {
-            unlink($folderPath);
-        }
-        $eqs = eqLogic::byType('qivivo');
-        foreach ($eqs as $eq) {
-            $eq->remove();
-        }
+	if ($pluginVersion < 2.0) {
+		//new custom API for new Comap interface:
+		$folderPath = dirname(__FILE__) . '/../../qivivo/exportedPrograms/';
+		if (is_dir($folderPath)) unlink($folderPath);
+		$eqs = eqLogic::byType('qivivo');
+		foreach ($eqs as $eq) {
+			$eq->remove();
+		}
 
-    }
+	}
 
-    //resave eqs for new cmd:
-    try
-    {
-        $eqs = eqLogic::byType('qivivo');
-        foreach ($eqs as $eq) {
-            $eq->save();
-        }
-    }
-    catch (Exception $e)
-    {
-        $e = print_r($e, 1);
-        log::add('qivivo', 'error', 'qivivo_update ERROR: '.$e);
-    }
+	//resave eqs for new cmd:
+	try
+	{
+		$eqs = eqLogic::byType('qivivo');
+		foreach ($eqs as $eq)
+		{
+			$eq->save();
+		}
+	}
+	catch (Exception $e)
+	{
+		$e = print_r($e, 1);
+		log::add('qivivo', 'error', 'qivivo_update ERROR: '.$e);
+	}
 
-    config::save('pluginversion', 2.1, 'qivivo');
+	config::save('pluginversion', 2.1, 'qivivo');
 }
 
-function qivivo_remove() {}
+function qivivo_remove() {
+
+}
 
 ?>
